Statistics update (#1587)
* updated method with month statistics to return dto

* minor improvements

* deleted unnecessary lombok annotations

+++ core/src/main/java/greencity/service/GraphService.java
@@ -17,5 +17,5 @@ public interface GraphService {
      *
      * @return {@link Map} with data for statistics.
      */
-    Map<Integer, Integer> getRegistrationStatistics();
+    Map<Integer, Long> getRegistrationStatistics();
 }
